Add proper online hub icon w/ unread notifications
